Attempt to limit setuptools for Snowflake snowpark (#57581)
Setuptools is used by many packages to build them but for snowpark
it seems that setuptools is used to "do stuff" as well. Snowpark
has a rather "low" expectations for setuptools, and setuptools
releases are happening quite often an there are many releases after
snowpark's minimum >=40.6.0 (released in 2018 and with way more
than 100 releases afterwards). This might be one of the reasons
that pip resolution hits "resolution too deep" error.

This PR adds lower-binding for setuptools to relatively new version
(April 2025 - more than 6 months ago) in an attempt to limit
the space of possible resolutions for `pip`. We also add the <9999
hack to signal pip that setuptools should be resolved early in
the process, which should also limit the potential space that pip
creates for resolution.