Improve tagging of providers when preparing a release (#57628)
Tagging providers after release after #57337 happened after preparing
the tarball - which was wrong :) as tarball uses the tag to build
the -sources package. Also it cannot be done too early becasue
tags are created based on the provider packages prepared in dist.

So - tagging providers *SHOULD* happen between creating provider
distributions and creating tarball.

This PR moves creation of the tarball to the right place, it also
improves a bit the "tag-providers" step:

* it uses PACKAGE_DATE environment variable if it is defined
* it does not react on existing tags - it simply overwrites the
  tags if such tags exist.