Replace the "postgres" extra on pandas from Presto provider with direct psycopg2-binary. (#57813)
This was pulling in psycopg2, but out own postres provider uses
psycopg2-binary which is essentially in conflict.

Since this wasn't behind an extra, and not every user needs it, _and_ since it
conflicts with our own postgres deps, I think removal is the best option.

This was only recently added/made working in #57141 (before then it issued a
warning about an unknown extra)

But some things have an implicit dep on Psycopg2 being available, so we
depend directly on the (correct for us) psycopg2-binary.