Improved how relative imports are allowed for shared distributions (#58824)
Previously we disallowed "parent" relative imports in shared
distributions and required to use TID252 noqa comment for those,
however this was not the best idea. It's ok to use `..` relative
imports in the shared distributions (to refer other shared
distributions). Also it's ok to use them inside the library,
and it's NOK to use them in `tests` for shared library.

This new way of allowing all kinds of relative rules for all shared
src/** is much more appropriate.